requirements.txt
PyYAML==3.10
pyparsing==2.0.1
